Create new Unit and UI test targets
